Squashed issue #9.
The feet and hands were being properly positioned, but the error checking was
checking the wrong values, so it seemed as if you were getting errors. Still,
the arms and legs don't necessarily go where you want them to machine
precision. I'm only getting mm accuracy on some of them. Something worth
checking in the future.