Add RawData files for the Balanceassistv1
Christoph and Jason measured the Balanceassistv1 bicycle, i.e. the first
balance assist bicycle we received from Bosch/Gazelle, with the standard
procedure. This is a modified Gazelle Grenoble step through frame with
the e-bike battery in the downtube. The torsional and compound pendulum
data files will be added to Figshare and the periods recorded in these
files result from that data and the current version of
BicycleParameters. The measurements were recorded mostly on July 30-1,
2024 and various corrections recorded upt to and including August 8,
2024.